step 41 , reducers
